Move interfaces for low-level readers to their own module.
